Major Logger Update Release
This update changes the logger process on all modules that log output. The previous logging required the logger instance to be passed to each function. The updated method does not need the logger to be passed but is defined in the main logger configuration. This update will break any previously called functions, so update functions before updating the module.